requirements.txt
requests
urllib3
Pillow
numpy